Add view for displaying single warehouses

+++ saleor/dashboard/warehouse/views.py
@@ -9,7 +9,11 @@ from django.utils.translation import pgettext_lazy
 
 from saleor.core.utils import get_paginator_items
 from saleor.dashboard.views import staff_member_required
-from saleor.dashboard.warehouse.forms import WarehouseForm
+from saleor.dashboard.warehouse.forms import (
+    WarehouseAddressForm,
+    WarehouseForm,
+    save_warehouse_from_forms,
+)
 from saleor.warehouse.models import Warehouse
 
 if TYPE_CHECKING:
@@ -19,47 +23,69 @@ if TYPE_CHECKING:
 
 @staff_member_required
 @permission_required("warehouse.manage_warehouses")
-def index(request: "HttpRequest"):
-    warehouses_qs = Warehouse.objects.all()
+def index(request: "HttpRequest") -> "HttpResponse":
+    warehouses_qs = Warehouse.objects.prefetch_data()
     warehouses = get_paginator_items(
         warehouses_qs, settings.DASHBOARD_PAGINATE_BY, request.GET.get("page")
     )
-    ctx = {"warehouses": warehouses}
+    ctx = {"warehouses": warehouses, "is_empty": not warehouses_qs.exists()}
     return TemplateResponse(request, "dashboard/warehouse/list.html", ctx)
 
 
 @staff_member_required
 @permission_required("warehouse.manage_warehouses")
 def warehouse_create(request: "HttpRequest") -> "HttpResponse":
-    form = WarehouseForm(request.POST or None)
-    if form.is_valid():
-        form.save()
+    warehouse_form = WarehouseForm(request.POST or None)
+    address_form = WarehouseAddressForm(request.POST or None)
+    if address_form.is_valid() and warehouse_form.is_valid():
+        warehouse = save_warehouse_from_forms(warehouse_form, address_form)
         msg = pgettext_lazy("Dashboard message", "Warehouse created")
         messages.success(request, msg)
-        return redirect("dashboard:warehouse-index")
-    ctx = {"form": form}
+        return redirect("dashboard:warehouse-detail", uuid=warehouse.uuid)
+    ctx = {"warehouse_form": warehouse_form, "address_form": address_form}
     return TemplateResponse(request, "dashboard/warehouse/form.html", ctx)
 
 
 @staff_member_required
 @permission_required("warehouse.manage_warehouses")
 def warehouse_update(request: "HttpRequest", uuid: "UUID") -> "HttpResponse":
-    warehouse = get_object_or_404(Warehouse, pk=uuid)
-    form = WarehouseForm(request.POST or None, instance=warehouse)
-    if form.is_valid():
-        form.save()
+    qs = Warehouse.objects.prefetch_data()
+    warehouse = get_object_or_404(qs, pk=uuid)
+    warehouse_form = WarehouseForm(request.POST or None, instance=warehouse)
+    address_form = WarehouseAddressForm(
+        request.POST or None, instance=warehouse.address
+    )
+    if address_form.is_valid() and warehouse_form.is_valid():
+        save_warehouse_from_forms(warehouse_form, address_form)
         msg = pgettext_lazy("Dashboard message", "Warehouse updated")
         messages.success(request, msg)
-        return redirect("dashboard:warehouse-index")
-    ctx = {"warehouse": warehouse, "form": form}
+        return redirect("dashboard:warehouse-detail", uuid=warehouse.uuid)
+    ctx = {
+        "warehouse": warehouse,
+        "warehouse_form": warehouse_form,
+        "address_form": address_form,
+    }
     return TemplateResponse(request, "dashboard/warehouse/form.html", ctx)
+
+
+@staff_member_required
+@permission_required("warehouse.manage_warehouses")
+def warehouse(request: "HttpRequest", uuid: "UUID") -> "HttpResponse":
+    qs = Warehouse.objects.prefetch_data()
+    ctx = {"warehouse": get_object_or_404(qs, pk=uuid)}
+    return TemplateResponse(request, "dashboard/warehouse/detail.html", ctx)
 
 
 @staff_member_required
 @permission_required("warehouse.manage_warehouses")
 def warehouse_delete(request: "HttpRequest", uuid: "UUID") -> "HttpResponse":
     warehouse = get_object_or_404(Warehouse, pk=uuid)
-    warehouse.delete()
-    msg = pgettext_lazy("Dashboard message", "Warehouse deleted")
-    messages.success(request, msg)
-    return redirect("dashboard-warehouse-index")
+    if request.method == "POST":
+        warehouse.delete()
+        msg = pgettext_lazy("Dashboard message", "Warehouse deleted")
+        messages.success(request, msg)
+        return redirect("dashboard:warehouse-list")
+    ctx = {"warehouse": warehouse}
+    return TemplateResponse(
+        request, "dashboard/warehouse/modal/confirm_delete.html", ctx
+    )
